--- docs/search-index.json.njk
@@ -1,6 +1,0 @@
----
-permalink: /search-index.json
----
-
-{{ collections.all | search | dump | safe }}
-
